Turn off lines-between class members & prefer default export
Turn off lines-between-class-members
This also gives a warning when there are no blank lines between property declarations in TS classes. TypeScript ESLint even recommends turning this (and all formatting rules) off in order to let a dedicated formatter decide that.

Turn of import/prefer-default-export
If a module is exporting more than one thing (e.g. an interface and a class), this will not warn. If the interface is later moved, it complains that a default export should be preferred. It does not make sense to have to update whether it's a default export depending on how many files it's exporting. It's also more convenient a lot of times to use named exports. E.g. when exporting re-exporting several things from an `index` file, we can simply do `export * from './thing'` instead of `export { default as thing } from './thing'.