Disable some rules
@typescript-eslint/no-explicit-any
There are cases where it's necessary to specify any, and `unknown` will not suffice. In general, `any` should be avoided, but I think it's a bit too strict to enforce it in all places, leading to a lot of unnecessary time spent somewhere where it doesn't really matter. There are active rules putting restrictions on variables with type `any` such as `no-unsafe-member-access` and `no-unsafe-call`.

no-param-reassign
There are cases where re-assigning parameters are helpful. E.g. to extract a piece of logic that might be duplicated, when the sole purpose of the function is to modify the parameter. Since we're not coding in a purely functional manner, this rule does not make too much sense.

promise/no-promise-in-callback
We should be allowed to call promises from inside e.g. forEach loop callbacks. The rule is not motivated in the rule's documentation.